requirements.txt
# OpenClaw Security Shield Dependencies
# Core
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.7
rich>=13.7.0
colorama>=0.4.6

# Security
cryptography>=41.0.7
pycryptodome>=3.20.0
requests>=2.31.0
urllib3>=2.1.0

# Code Analysis
astroid>=3.0.2
bandit>=1.7.6
safety>=2.3.5
pip-audit>=2.6.3

# Network Monitoring
psutil>=5.9.7
netifaces>=0.11.0
scapy>=2.5.0

# Logging & Audit
loguru>=0.7.2
python-json-logger>=2.0.7

# Database
sqlite3-api>=0.0.2

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1

# Async Support
aiohttp>=3.9.1
asyncio>=3.4.3

# Configuration Validation
pydantic>=2.5.3
jsonschema>=4.20.0

# Watchdog for File Monitoring
watchdog>=3.0.0
